fix: share homebrew variables across plays
I've tried 50 different ways to reuse the homebrew module
implementation. Adding the role under 'roles' in every play didn't work
because the variables would persist between plays. Maybe that could be
solved by adding the 'allow_duplicates' option to the role's meta.yml,
but in any case, calling homebrew for every play is not the best
performance wise.

The current solution works by introducing another playbook for homebrew.
Now, to my great frustration, variables were not persistent between
plays, because they did not share the homebrew role anymore. I solved
this by mapping variables to a fact in every play:

- set_fact:
    brew: "{{ brew }}"

Facts are host specific and persist for the entire site.yml playbook.
Together with the ansible.cfg 'hash_behaviour = merge', it allows me to
create a variable 'brew' containing all the necessary information. The
homebrew playbook comes last, and installs all dependencies defined in
the previous plays. This breaks if a brew formula is required to be
installed mid provisioning, but I'll deal with that when it happens.

Another solution I tried was having a single playbook and listing all
roles there:

This breaks if a brew formula is required to be installed mid
provisioning, but I'll deal with that when it happens.

Another solution I tried was having a single playbook and listing all
roles there:

This breaks if a brew formula is required to be installed mid
provisioning, but I'll deal with that when it happens.

Another solution I tried was having a single playbook and listing all
roles there:

- name: "=== all ==="
  hosts: machine
  roles:
    - zsh
    - alacritty
    - vim
    - homebrew

This achieves the same result, without the use of the fact 'hack', but
with reduced granularity.