feat(provision): add 'setup_homebrew' tag
The 'do_homebrew' tag was used for two purposes:
- running tasks for setting up homebrew (using role)
- installing formulae/casks/taps (using custom tasks)

Now there are separate tags for these actions. This allows us to use the
dotfiles cli to install brew resources for one or many topics without
the overhead of making sure homebrew is installed (it should be).

The bootstrap script has been updated to include the 'setup_homebrew'
tag.