refactor(provision): revamp topic_tasks
BREAKING CHANGE: removes the 'do_post_provision' tag.

Tasks kept in `topic.tasks.yaml` files, previously known as
'post_provision_tasks' are now refered to as 'topic_tasks'. They are
executed immediately after the files are sourced, i.e. in the
eliasnorrby/dotfiles role, rather than as the last step of the playbook.

Tasks are only run if their corresponding topic is enabled. The check
for this is rather cumbersome, but is the best I could find at this
time.

```
 # breaks if shell/zsh is undefined
when: topics.shell.zsh.state == "present"
 # requires installing jmespath
when: topics | json_query('shell.zsh.state') == "present"
 # works
when:
  - topics.shell.zsh is defined
  - topics.shell.zsh.state == "present"
```

closes: #20