fix(provision): properly filter topic tasks based
A solution for only running topic tasks for present topics was
introduced in

6955cf7 refactor(provision): revamp topic_tasks

but with some errors. The method for checking whether a particular topic
is present was flawed.

when:
 - topics.editor.vscode is defined
 - tpoics.editor.vscode.state == 'present'

The first condition will never be true, because topics.editor is a list,
not a dictionary.

This commit moves to filtering to the top level step where task files
are included, removing the need for having individual when clauses in
the topic task files themselves.

It works by extracting the topic path property from the actual path of
each task file, and only including the tasks contained in each
respective file if there's an entry in present_topics with a matching
value for config.path.