feat(bootstrap): check for executables before install
This will (hopefully) make the build pass, but it will prove problematic
in a real scenario. I can upgrade openssl and python on Travis before
running the script, but I won't be on a fresh MacBook. In that case I'll
likely hit the same certificate error when running ansible-galaxy
install. Maybe the one-line bootstrap dream is over...