feat(vim): use unnamedplus for clipboard
On Windows and MacOS, the * and + registers both point to the system
clipboard. On Linux, there's also the selection clipboard (mapped to
register *). Using only unnamedplus allows us to yank and paste in with,
and copy/paste using ctrl/cmd + c/v in other windows, on all platforms.

More info: https://stackoverflow.com/a/30691754