feat(xkb): map right alt to hyper key (mod3)
Well, this was a good to days' worth of tinkering. I've been down
rabbit holes all over the place, reading up on xmodmap, setxkbmap, their
individual config files and command line options; xcape; xev; xdotool;
and more.

Left Alt and Left Win are now swapped. So, the Super key is where the
command key is on Apple keyboards. Caps Lock remains mapped to Ctrl when
held and Escape when tapped. Right Alt is mapped to Hyper. It's done in
a most peculiar way, and I don't know how maintainable it is. We will
have to see.

sxhkd does not seem happy about reloading its config (pkill -USR1 -x
sxhkd), at least not after having re-run setup-keybindings.sh. The
most consistent way to refresh it seems to be restarting bspwm, which
will in turn set up the bindings and restart sxhkd.