feat(polybar): try to make tray behave
The dropbox icon is problematic. Sometimes it shows up on monitor 1,
sometimes on monitor 2. When it renders, there are artifacts from the
previous background behind it. Some suggest adding a 'padding' module
with empty text to get around this. Not very happy with the result, but
it will do for now.