feat(tmux): skip the initial rename session prompt
The hook on after-new-session asks for a session name, but this is
unnecessary on the first launch.

This skip is a little hacky: wait for a fraction of a second and send a
return key press to the window using xdotool, but I've spent enough time
look for a way to bypass tmux hooks that I'm happy with it.

(What will I do when Wayland comes around?)