refactor(bspwm): generalize spotify workaround
It seems the issues with bspc rules and spotify is general to all
electron apps (including, e.g. discord). I generalized the wrapper
script to take a desktop and an executable (maybe even args in the
future...) as arguments. That had me update the focus_or_run script to
make it pass any extra arguments on the executable it tries to run.

What was

  focus_or_run executable class

is now

  focus_or_run class executable [arbitrary list of arguments]

e.g.

  focus_or_run spotify launch_on_desktop 7 spotify