feat(vim): update project-wide search functions
I become bothered by the :Rg command matching filenames in the list of
search results returned by ripgrep. It has some uses, but most of the
time it makes it harder to find a good match (try searching for
'keybindings' in this repo for example).

I've set up three mappings calling different variants of the Rg command:

<leader>// : ripgrep results filtered by fzf, not matching filenames
<leader>/? : reload ripgrep results when query is updated
<leader>/f : call the standard Rg function (with filename matching)

Resources:
- https://github.com/junegunn/fzf/issues/1109#issuecomment-339983902
- https://github.com/junegunn/fzf.vim#example-rg-command-with-preview-window
- https://github.com/junegunn/fzf.vim#example-advanced-ripgrep-integration
- https://github.com/junegunn/fzf.vim/issues/346#issuecomment-288483704