feat(zsh): switch yq implementation on arch
I discovered there are (at least) two projects under the name yq:

mikefarah/yq
kislyuk/yq

the latter of which seems to be more of a pure wrapper around jq. I
don't really know which is better, but mikefarah's implementation is
what is available on homebrew as yq, so to keep things consistent,
that's what I plan on using on arch as well.