feat(arch): handle focus of withdrawn windows
There are so many special snowflakes when it comes to windows. Some
behave like you expect: app running - a window to focus; quit the app -
no windows left.

Chromium has many windows, only one of which can be focused. We have to
find the correct one to call xdotool windowactivate on. But at least
when it's quit, the windows go away.

1Password will leave a bunch of windows in the background, none of which
can be activated. The best way to create a new window is to 'exec
1password'.

Slack's window will remain when quitting, but with a 'Withdrawn' state.
This means it will pass the check for a suitable window to activate, but
xdotool windowactivate will do nothing. The best way I've found to bring
it to the foreground is to 'exec slack'.

Also: zsh does not do word splitting like bash or sh.

  wids=$(xdotool search --class 'chromium')
  for wid in $wids; do
    echo "Window: $wid"
  done

will work wonders in the latter, but only result in one iteration in the
former. However,

  for wid in $(xdotool search --class 'chromium'); do
    echo "Window: $wid"
  done

will work in all three. In zsh, one can request word splitting using
${=var}:

  wids=$(xdotool search --class 'chromium')
  for wid in ${=wids}; do
    echo "Window: $wid"
  done