feat(zsh): add man_widget
One full day's work on a fun little zle widget. It started out as a
solution to a recurring annoyance I face along the lines of 'what does
that flag do to that command again?' E.g., I want to know what

  curl -sSLO 'url...'

does before running it, or why somebody used '-n' or '-e' in their sed
command.

To find the answer, I would browse the command manpage or --help.
Searching for a flag is not as simple as typing

  /-s

though. To find the correct session, I need something like

  /^[[:blank]]*-s

which is tedious to write.

My idea was to create a script around fzf, providing a fuzzy finding
interface to a command's options, with the relevant section of the docs
displayed in the preview window.

While working, this evolved into a tool to help both with documentation
lookup, but also command editing.

I've created a zle widget, bound to ^F^M (for 'fuzzy manpage'). It
operates in two modes: 'lookup', and 'complete'.

The lookup mode is invoked by pressing ^F^M with an empty buffer.
Executable commands (from 'compgen -c') are filtered with fzf. On
choosing one, another fzf session is invoked, now reading 'man
<command>' or '<command> --help' and offering up a list of flags that
can be passed. A preview script using awk attempts to extract the
relevant parts of the docs. On hitting return, less is launched with the
complete 'help input', be it from 'man' or from '--help', with a search
performed for the flag. That is, I'm taken to the relevant section in
the full docs, free to browse the surrounding sections.

The complete mode is invoked by pressing ^F^M with a non-empty buffer. I
try to collect the relevant command from the buffer contents to figure
out which command to complete. Then the same flag-filtering-session as
in the lookup mode is invoked, but on selection, the flag is inserted on
the command line instead of jumping to the docs.

For example:

  $ find ^F^M

invokes an fzf session with the flags for find, naturally, but I wanted
to handle pipes and sub-shells to some extent as well, so

  $ find . -type f | grep ^F^M

brings up flags for grep, while

  $ cat $(ls ^F^M

brings up flags for ls. The cursor position is respected, so

  $ kafka-topics --list ^F^M | grep 'some-prefix'

would bring up flags for kafka-topics, restoring the cursor position
afterwards.

Something I want to work on next is supporting chained commands, e.g.

  $ git commit
  $ docker run
  $ gh issue create

As of now, invoking the widget with these commands in the buffer would
source documentation only for the base command, whereas what would be
more appropriate would be to take the entire change add append '--help'
to that.