fix(zsh): handle local scripts with man_widget
If one calls man on a file like this:

  $ man ./deploy.sh

man will page the entire contents of the file. This is probably not what
we want, but this means it will pass the has-man check in man_widget.
Therefore, if the command begins with a dot '.', fail the test and
resort to --help.