fix(zsh): add _os conditional around compgen
compgen is a bash builtin - it's not available in zsh. On MacOS, there
seems to be an emulation in place (you can view the function definition
with `which compgen`). I've yet to figure out where it's sourced, and I
don't have time to look into it now.

On arch, fall back to bash -c 'compgen -c' to list available programs.
There may be discrepancies in how $PATH is constructed, but it seems to
work okay.