fix(zsh): only remove space if expanding
If we define a blank alias like

  balias clh="curl localhost:"

we cannot type e.g.

  echo clh<space>

Since the alias is not global, it won't be expanded, so there will be no
trailing space added. Instead, the space we're trying to insert will be
deleted. This fix assumes a blank alias will only be expanded if it's
typed at the beginning of the line, and avoids setting 'remove_space'
otherwise.