feat(sxhkd): add arrow key hyper bindings
I've worked so long to achieve this, and tried so many different
solutions, that seeing how few lines it took feels strange.

It seems like such a simple thing: I want to send a key press to the
focused window.

Sending keystrokes to _other_ windows is trivial. I can freely move the
cursor in another window by specifying its id:

  hyper + j
    xdotool key --window <wid> Down

So why won't this work for the focused window?

  hyper + j
    xdotool key Down

The thing is (I'm only guessing a little here), that's exactly what
sxhkd is set up to monitor: key presses targeting the focused window.
Those are the keys that should trigger bindings, so they are intercepted
by sxhkd.

One solution is to prefix the command with a sleep call:

  hyper + j
    sleep 0.2; xdotool key Down

And while this works, it is a sluggish experience at best, and trying to
fine tune the duration leads to inconsistent behaviour.

What I found was that I could set up another set of bindings to trigger
on the key presses generated by the first. Using the ~ prefix, the
captured event is replayed for other clients, i.e. passed through to the
window.

  hyper + {h,j,k,l}
    xdotool key {Left,Down,Up,Right}

  hyper + ~{Left,Down,Up,Right}
    # no-op

Hooray!