feat(xkb): add workman layout & refactor xkb
Rework how keyboard layouts are set. Instead of setting options with
setxkbmap and piping the output of setxkbmap -print to xkbcomp, call
xkbcomp directly, pointing to a dedicated config file. The hyper key
customizations live (mostly) in layout.xkb now, together with the
layouts.

Furthermore, use ctrl:nocaps instead of capslock:ctrl_modifier.

But most importantly: add the Workman-P layout, slight modifications.
I'm not ready to replace my Caps Lock / Escape binding for another
backspace, and I need to ignore the ralt_switch to get the hyper key to
work.

Modifying the xmodmap call with repeated Hyper_L's seems to be required
for making the binding work in all layouts.

Weirdly enough, sxhkd does not seem to react to layouts being swapped
(no matter whether it's launched with the -m -1 option). My hyper
bindings still work as if I'm using a qwerty layout. Initially, this may
be helpful - I can concentrate on learning to type with workman, while
using familiar navigation, but in the long run, it'll just troublesome.

I tried (and failed) to modify the evdev.xml and evdev.lst files in
/usr/share/X11/xkb to get the polybar xkeyboard module to show different
names for the different variants. As of now, it shows 'English (US)' for
both. There's an unreleased %variant% token I can use, hopefully when
version 3.6.0 is released.

Resources:
- http://linux.lsdev.sil.org/wiki/index.php/Building_an_XKB_Keyboard
- https://github.com/workman-layout/Workman/tree/master/xorg
- https://hack.org/mc/writings/xkb.html
- https://wiki.archlinux.org/title/X_keyboard_extension