fix(ssh): add alias with TERM set
TIL: The TERM variable is passed when making an ssh connection.
tmux-256color is not a widely known value (and neither is 'alacritty', I
presume). This leads to improper handling of some things, i.e. opening
vim in a remote session.

One could potentially use the SetEnv directive instead of an alias like
this, but that requires the host to be configured to accept the TERM
environment variable.