feat(vim): add method reference mapping
This is only for typescript, so should likely be moved from core. But
I've had issues with setting up keymaps in ftplugin files.

One-shot using Gemini:

```ts
export class MyClass {
  constructor() {
    // Initialization code here
  }

  myMethod() {
    // Method implementation here
  }
}
```

I'm working in a typescript project. I often want to refer to different
service methods in writing. I like to do this using this syntax:
`MyClass#myMethod`.

I'd like a vim binding that allows me to quickly copy a snippet like
that. I think we should be able to use tree-sitter to parse the
TypeScript code and extract the class name and method name.

Pseudo code:
- given the cursor position
- use tree-sitter to get containing method, if any
- get containing class, if any
- format the string as `ClassName#methodName`
- print error message if none found