chore(cicd): remove tekton and argocd dependencies
Haven't used them in ages. I wonder what I should do when a topic grows
in a way where I only end up using a subset of its dependencies. Can't
disable it, ought to split it up I guess. But in cases when there's no
config for the tools, I find it's just easier to remove the
dependencies.